+++ carp_mobile_sensing/lib/domain/domain.dart
@@ -1,6 +1,6 @@
 /// The core CARP domain classes like [Study], [Task], and [Measure].
 /// Also hold JSON serialization logic to serialize [Datum] objects
-/// into a [CARPDataPoint] as well as deseralization of [Study] objects
+/// into a [DataPoint] as well as deseralization of [Study] objects
 /// obtained from a [StudyManager].
 library domain;
 
@@ -16,7 +16,7 @@ import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 part 'serialization.dart';
 part 'study.dart';
 part 'tasks.dart';
-part 'carp_datapoint.dart';
+part 'datapoint.dart';
 part 'datum.dart';
 part 'measures.dart';
 part 'device_info.dart';