+++ carp_mobile_sensing/test/study_json_test.dart
@@ -7,9 +7,8 @@ String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(objec
 
 void main() {
   setUp(() {
-    //SamplingPackageRegistry.register(AudioSamplingPackage());
-    //SamplingPackageRegistry.register(CommunicationSamplingPackage());
-    //SamplingPackageRegistry.register(ContextSamplingPackage());
+    // This is a hack. Need to create some serialization object in order to intialize searialization.
+    Study study = Study("123", "kkk");
   });
 
   /// Test if we can load a raw JSON from a file and convert it into a [Study] object with all its [Task]s and [Measure]s.