+++ carp_webservices/lib/carp_service/carp_service.dart
@@ -16,13 +16,16 @@ import 'dart:io';
 import 'package:carp_webservices/carp_auth/carp_auth.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'dart:math';
+import 'package:json_annotation/json_annotation.dart';
 
+part 'carp_datapoint.dart';
 part 'carp_app.dart';
 part 'datapoint_reference.dart';
 part 'file_reference.dart';
 part 'document_reference.dart';
 part 'carp_tasks.dart';
 part 'push_id_generator.dart';
+part 'carp_service.g.dart';
 
 /// Provide access to the CARP web services endpoint.
 ///
@@ -151,11 +154,11 @@ class CarpService {
   /// [id] can be omitted if a local file is not uploaded yet.
   FileStorageReference getFileStorageReference([int id]) => FileStorageReference._(this, id);
 
-//  /// Gets a [DocumentReference] for the specified unique id.
-//  DocumentReference document(int id) {
-//    assert(id != null);
-//    return DocumentReference._id(this, id);
-//  }
+  /// Gets a [DocumentReference] for the specified unique id.
+  DocumentReference documentById(int id) {
+    assert(id != null);
+    return DocumentReference._id(this, id);
+  }
 
   /// Gets a [DocumentReference] for the specified CARP Service path.
   DocumentReference document(String path) {