+++ carp_webservices/lib/carp_service/datapoint_reference.dart
@@ -146,3 +146,5 @@ class DataPointReference extends CarpReference {
     }
   }
 }
+
+String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(object);