+++ carp_backend/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.1
+* adjustment to new CARP web service API as implemented in `carp_webservices` package.
+
 ## 0.3.0
 * Adjustment to refactoring of `carp_mobile_sensing` library to version 0.3.0