+++ carp_webservices/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.1
+* support for get document by id
+
 ## 0.2.0
 * upgrade and test on Flutter v. 1.3.4 Dart v. 2.2.1 
 * upgrading to the new CARP Web Service API for documents (instead of objects) and collections